+++ apps/backoffice-v2/src/domains/entities/hooks/useSelectEntityFilterOnMount/useSelectEntityFilterOnMount.tsx
@@ -1,13 +1,14 @@
 import { useFiltersQuery } from '../../../filters/hooks/queries/useFiltersQuery/useFiltersQuery';
 import { useEffect, useMemo } from 'react';
 import { useSearchParamsByEntity } from '../../../../common/hooks/useSearchParamsByEntity/useSearchParamsByEntity';
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { useEntityType } from '../../../../common/hooks/useEntityType/useEntityType';
 import { searchParamsToObject } from '../../../../common/hooks/useZodSearchParams/utils/search-params-to-object';
+import { useLocale } from '@/common/hooks/useLocale/useLocale';
 
 export const useSelectEntityFilterOnMount = () => {
   const { data: filters } = useFiltersQuery();
-  const { locale } = useParams();
+  const locale = useLocale();
   const [{ filterId }, setSearchParams] = useSearchParamsByEntity();
   const entity = useEntityType();
   const navigate = useNavigate();