+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/common
 
+## 0.9.10
+
+### Patch Changes
+
+- document changes
+
+## 0.9.9
+
+### Patch Changes
+
+- Added a sort direction type
+
 ## 0.9.8
 
 ### Patch Changes